class as separate module
